Move core Kotlin files into conventional directories, and add some really sucky tests.
