+++ core/src/main/kotlin/org/eln2/math/FunctionTable.kt
