+++ core/src/main/kotlin/org/eln2/math/IFunction.kt
