+++ core/src/main/kotlin/org/eln2/serialization/generic/IHaveState.kt
