+++ core/src/main/kotlin/org/eln2/serialization/generic/ISerializeLazily.kt
