+++ core/src/main/kotlin/org/eln2/serialization/generic/JsonSerializer.kt
