+++ core/src/main/kotlin/org/eln2/sim/IProcess.kt
