+++ core/src/main/kotlin/org/eln2/sim/electrical/mna/Circuit.kt
@@ -35,11 +35,11 @@ class Circuit {
     val postProcess = WeakHashMap<IProcess, Unit>()
 
     // These fields are only for the solver--don't use them casually elsewhere
-    private var matrix: RealMatrix? = null
-    private var knowns: RealVector? = null
-    private var solver: DecompositionSolver? = null
-    private var nodes: List<WeakReference<Node>> = emptyList()
-    private var vsources: List<WeakReference<VSource>> = emptyList()
+    internal var matrix: RealMatrix? = null
+    internal var knowns: RealVector? = null
+    internal var solver: DecompositionSolver? = null
+    internal var nodes: List<WeakReference<Node>> = emptyList()
+    internal var vsources: List<WeakReference<VSource>> = emptyList()
 
     /* From Falstad: declare that a potential change dV in node b changes the current in node a by x*dV, complicated
        slightly by independent voltage sources. The unit of x is "Mhos", reciprocal Ohms, a unit of conductance.
@@ -214,130 +214,5 @@ class Circuit {
 
         components.forEach { it.postStep(dt) }
         postProcess.keys.forEach { it.process(dt) }
-    }
-
-    companion object {
-        @JvmStatic
-        fun main(args: Array<String>) {
-            main_1()
-            println("---")
-            main_2()
-            println("---")
-            main_3()
-            println("---")
-        }
-        
-        fun main_1() {
-            println("main_1: === Basic circuit test ===")
-            val c = Circuit()
-
-            val vs = VoltageSource()
-            val r1 = Resistor()
-
-            c.add(vs)
-            c.add(r1)
-
-            vs.connect(1, r1, 0)
-            vs.connect(0, r1, 1)
-            vs.connect(1, c.ground)
-
-            vs.u = 10.0
-            r1.r = 100.0
-
-            c.step(0.5)
-
-            print("main_1: matrix:\n${MATRIX_FORMAT.format(c.matrix)}")
-
-            for(comp in c.components) {
-                println("main_1: comp $comp name ${comp.name} nodes ${comp.nodes} vs ${comp.vsources}")
-                for(node in comp.nodes) {
-                    println("\t node $node index ${node.node.index} ground ${node.node.isGround} potential ${node.node.potential}")
-                }
-            }
-
-            for(node in c.nodes) {
-                println("main_1: node ${node.get()} index ${node.get()?.index} potential ${node.get()?.potential}")
-            }
-
-            println("main_1: vs current: ${vs.i}\nmain_1: r1 current: ${r1.i}")
-        }
-        
-        fun main_2() {
-            println("main_2: === Capacitors and Inductors ===")
-            val c = Circuit()
-
-            val vs = VoltageSource()
-            val c1 = Capacitor()
-            val l1 = Inductor()
-            val r1 = Resistor()
-            val r2 = Resistor()
-
-            c.add(vs, c1, l1, r1, r2)
-
-            vs.connect(1, c.ground)
-            vs.connect(0, r1, 0)
-            vs.connect(0, r2, 0)
-            r1.connect(1, c1, 0)
-            r2.connect(1, l1, 0)
-            c1.connect(1, c.ground)
-            l1.connect(1, c.ground)
-
-            vs.u = 10.0
-            r1.r = 10.0
-            r2.r = 10.0
-            c1.c = 0.01
-            l1.h = 1.0
-
-            val fp = PrintStream(FileOutputStream("main_2.dat"))
-            fp.println("#t\tc1.i\tl1.i\tr1.u\tr2.u\tvs.i")
-            var t = 0.0
-            val st = 0.05
-            for(i in 0 until 25) {
-                c.step(st)
-                t += st
-                println("main_2: t=$t c1.i=${c1.i} l1.i=${l1.i} r1(c1).u=${r1.u} r2(l1).u=${r2.u} vs.i=${vs.i}")
-                fp.println("$t\t${c1.i}\t${l1.i}\t${r1.u}\t${r2.u}\t${vs.i}")
-            }
-            fp.close()
-        }
-
-        fun main_3() {
-            println("main_3: === Diodes ===")
-            val c = Circuit()
-
-            val vs = VoltageSource()
-            val r1 = Resistor()
-            val r2 = Resistor()
-            val r3 = Resistor()
-            val d1 = RealisticDiode(DiodeData.default)
-            val d2 = RealisticDiode(DiodeData.diodes["falstad-zener"] ?: error("no zener"))
-            val d3 = IdealDiode()
-            val diodes = arrayOf(d1, d2, d3)
-
-            c.add(vs, r1, r2, r3, d1, d2, d3)
-            
-            vs.connect(1, c.ground)
-            for(diode in diodes)
-                vs.connect(0, diode, 0)
-            for((r, d) in arrayOf(Pair(r1, d1), Pair(r2, d2), Pair(r3, d3))) {
-                d.connect(1, r, 0)
-                r.connect(1, c.ground)
-                r.r = 10.0
-            }
-
-            val fp = PrintStream(FileOutputStream("main_3.dat"))
-            fp.println("#t\tvs.u\tr1.i\tr1.p\tr2.1\tr2.p\tr3.i\tr3.p")
-            var t = 0.0
-            val st = 0.05
-            for(i in -10 .. 10) {
-                vs.u = i.toDouble()
-                c.step(st)
-                if(i == -10) println("main_3: matrix: ${MATRIX_FORMAT.format(c.matrix)}")
-                t += st
-                println("main_3: t=$t vs.u=${vs.u} r1.i=${r1.i} r1.p=${r1.p} r2.i=${r2.i} r2.p=${r2.p} r3.i=${r3.i} r3.p=${r3.p}")
-                println("... knowns=${c.knowns}")
-                fp.println("$t\t${vs.u}\t${r1.i}\t${r1.p}\t${r2.i}\t${r2.p}\t${r3.i}\t${r3.p}")
-            }
-        }
     }
 }
