+++ core/src/main/kotlin/org/eln2/sim/electrical/mna/component/Resistor.kt
