+++ core/src/main/kotlin/org/eln2/space/Graph.kt
