+++ core/src/test/kotlin/org/eln2/sim/electrical/mna/CircuitTest.kt
@@ -1,0 +1,131 @@
+package org.eln2.sim.electrical.mna
+
+import org.eln2.sim.electrical.mna.component.*
+import org.junit.jupiter.api.Assertions
+import org.junit.jupiter.api.Test
+import java.io.*
+
+internal class CircuitTest {
+
+    @Test
+    fun basicCircuitTest() {
+        val c = Circuit()
+
+        val vs = VoltageSource()
+        val r1 = Resistor()
+
+        c.add(vs)
+        c.add(r1)
+
+        vs.connect(1, r1, 0)
+        vs.connect(0, r1, 1)
+        vs.connect(1, c.ground)
+
+        vs.u = 10.0
+        r1.r = 100.0
+
+        c.step(0.5)
+
+        print("main_1: matrix:\n${MATRIX_FORMAT.format(c.matrix)}")
+
+        for(comp in c.components) {
+            println("main_1: comp $comp name ${comp.name} nodes ${comp.nodes} vs ${comp.vsources}")
+            for(node in comp.nodes) {
+                println("\t node $node index ${node.node.index} ground ${node.node.isGround} potential ${node.node.potential}")
+            }
+        }
+
+        for(node in c.nodes) {
+            println("main_1: node ${node.get()} index ${node.get()?.index} potential ${node.get()?.potential}")
+        }
+
+        println("main_1: vs current: ${vs.i}\nmain_1: r1 current: ${r1.i}")
+    }
+
+    @Test
+    fun capacitorsAndInductors() {
+        val c = Circuit()
+
+        val vs = VoltageSource()
+        val c1 = Capacitor()
+        val l1 = Inductor()
+        val r1 = Resistor()
+        val r2 = Resistor()
+
+        c.add(vs, c1, l1, r1, r2)
+
+        vs.connect(1, c.ground)
+        vs.connect(0, r1, 0)
+        vs.connect(0, r2, 0)
+        r1.connect(1, c1, 0)
+        r2.connect(1, l1, 0)
+        c1.connect(1, c.ground)
+        l1.connect(1, c.ground)
+
+        vs.u = 10.0
+        r1.r = 10.0
+        r2.r = 10.0
+        c1.c = 0.01
+        l1.h = 1.0
+
+        val actual = ByteArrayOutputStream()
+        val fp = PrintStream(actual)
+
+        fp.println("#t\tc1.i\tl1.i\tr1.u\tr2.u\tvs.i")
+        var t = 0.0
+        val st = 0.05
+        for(i in 0 until 25) {
+            c.step(st)
+            t += st
+            println("main_2: t=$t c1.i=${c1.i} l1.i=${l1.i} r1(c1).u=${r1.u} r2(l1).u=${r2.u} vs.i=${vs.i}")
+            fp.println("$t\t${c1.i}\t${l1.i}\t${r1.u}\t${r2.u}\t${vs.i}")
+        }
+        fp.close()
+
+        val expected = FileInputStream("testdata/main_2.dat").readAllBytes()
+        org.junit.jupiter.api.Assertions.assertArrayEquals(expected, actual.toByteArray())
+    }
+
+    @Test
+    fun diodes() {
+        val c = Circuit()
+
+        val vs = VoltageSource()
+        val r1 = Resistor()
+        val r2 = Resistor()
+        val r3 = Resistor()
+        val d1 = RealisticDiode(DiodeData.default)
+        val d2 = RealisticDiode(DiodeData.diodes["falstad-zener"] ?: error("no zener"))
+        val d3 = IdealDiode()
+        val diodes = arrayOf(d1, d2, d3)
+
+        c.add(vs, r1, r2, r3, d1, d2, d3)
+
+        vs.connect(1, c.ground)
+        for(diode in diodes)
+            vs.connect(0, diode, 0)
+        for((r, d) in arrayOf(Pair(r1, d1), Pair(r2, d2), Pair(r3, d3))) {
+            d.connect(1, r, 0)
+            r.connect(1, c.ground)
+            r.r = 10.0
+        }
+
+        val actual = ByteArrayOutputStream()
+        val fp = PrintStream(actual)
+        fp.println("#t\tvs.u\tr1.i\tr1.p\tr2.1\tr2.p\tr3.i\tr3.p")
+        var t = 0.0
+        val st = 0.05
+        for(i in -10 .. 10) {
+            vs.u = i.toDouble()
+            c.step(st)
+            if(i == -10) println("main_3: matrix: ${MATRIX_FORMAT.format(c.matrix)}")
+            t += st
+            println("main_3: t=$t vs.u=${vs.u} r1.i=${r1.i} r1.p=${r1.p} r2.i=${r2.i} r2.p=${r2.p} r3.i=${r3.i} r3.p=${r3.p}")
+            println("... knowns=${c.knowns}")
+            fp.println("$t\t${vs.u}\t${r1.i}\t${r1.p}\t${r2.i}\t${r2.p}\t${r3.i}\t${r3.p}")
+        }
+
+        val expected = FileInputStream("testdata/main_3.dat").readAllBytes()
+        Assertions.assertArrayEquals(expected, actual.toByteArray())
+    }
+}
