+++ core/testdata/main_2.dat
@@ -1,0 +1,26 @@
+#t	c1.i	l1.i	r1.u	r2.u	vs.i
+0.05	-0.0	0.3333333333333333	6.666666666666667	3.333333333333334	-1.0
+0.1	-0.6666666666666666	0.5555555555555556	4.4444444444444455	5.555555555555555	-1.0000000000000002
+0.15000000000000002	-1.111111111111111	0.7037037037037037	2.9629629629629655	7.037037037037038	-1.0000000000000002
+0.2	-1.407407407407407	0.8024691358024691	1.975308641975312	8.024691358024691	-1.0000000000000002
+0.25	-1.6049382716049376	0.8683127572016461	1.3168724279835438	8.683127572016462	-1.0000000000000002
+0.3	-1.7366255144032912	0.9122085048010974	0.8779149519890304	9.122085048010975	-1.0000000000000004
+0.35	-1.8244170096021939	0.9414723365340649	0.5852766346593548	9.414723365340649	-1.0000000000000002
+0.39999999999999997	-1.882944673068129	0.9609815576893767	0.3901844231062377	9.609815576893766	-1.0000000000000002
+0.44999999999999996	-1.9219631153787524	0.9739877051262511	0.260122948737493	9.73987705126251	-1.0000000000000002
+0.49999999999999994	-1.9479754102525013	0.9826584700841674	0.17341529915833043	9.826584700841675	-1.0000000000000004
+0.5499999999999999	-1.965316940168334	0.9884389800561116	0.11561019943888873	9.884389800561117	-1.0000000000000002
+0.6	-1.9768779601122222	0.9922926533707411	0.07707346629259426	9.922926533707411	-1.0000000000000004
+0.65	-1.9845853067414811	0.9948617689138274	0.051382310861731284	9.948617689138274	-1.0000000000000004
+0.7000000000000001	-1.9897235378276537	0.9965745126092183	0.034254873907823224	9.965745126092182	-1.0000000000000004
+0.7500000000000001	-1.9931490252184354	0.9977163417394789	0.02283658260521726	9.977163417394788	-1.0000000000000004
+0.8000000000000002	-1.9954326834789566	0.9984775611596526	0.01522438840347995	9.984775611596525	-1.0000000000000004
+0.8500000000000002	-1.996955122319304	0.9989850407731017	0.010149592268987817	9.989850407731018	-1.0000000000000007
+0.9000000000000002	-1.9979700815462025	0.9993233605154012	0.006766394845993062	9.99323360515401	-1.0000000000000004
+0.9500000000000003	-1.9986467210308014	0.9995489070102674	0.0045109298973304846	9.995489070102675	-1.0000000000000004
+1.0000000000000002	-1.999097814020534	0.9996992713401783	0.003007286598220915	9.996992713401783	-1.0000000000000002
+1.0500000000000003	-1.9993985426803558	0.9997995142267856	0.002004857732147869	9.997995142267856	-1.0000000000000002
+1.1000000000000003	-1.9995990284535705	0.999866342817857	0.0013365718214330968	9.99866342817857	-1.0000000000000004
+1.1500000000000004	-1.9997326856357134	0.9999108952119047	8.910478809571742E-4	9.999108952119046	-1.0000000000000002
+1.2000000000000004	-1.9998217904238085	0.9999405968079365	5.940319206398925E-4	9.999405968079365	-1.0000000000000004
+1.2500000000000004	-1.999881193615872	0.9999603978719577	3.9602128042837137E-4	9.999603978719577	-1.0000000000000002
