+++ core/testdata/main_3.dat
@@ -1,0 +1,22 @@
+#t	vs.u	r1.i	r1.p	r2.1	r2.p	r3.i	r3.p
+0.05	-10.0	-1.714352821952828E-7	2.939005598137625E-13	-0.4531330251367563	2.053295384695882	-9.999999989999993E-10	9.999999979999986E-18
+0.1	-9.0	-1.7143528189869866E-7	2.9390055879686276E-13	-0.35313445393772996	1.2470394255789872	-8.999999990999994E-10	8.099999983799988E-18
+0.15000000000000002	-8.0	-1.714352818986986E-7	2.939005587968626E-13	-0.2531358827387038	0.640777751299028	-7.999999991999994E-10	6.399999987199991E-18
+0.2	-7.0	-1.7143528189869863E-7	2.9390055879686266E-13	-0.15313731153967752	0.23451036185600252	-6.999999992999995E-10	4.899999990199993E-18
+0.25	-6.0	-1.7143528189869866E-7	2.9390055879686276E-13	-0.05313874034065127	0.028237257249911586	-5.999999993999996E-10	3.599999992799995E-18
+0.3	-5.0	-1.7143528189869863E-7	2.9390055879686266E-13	0.046859830858374994	0.021958437480755134	-4.999999994999995E-10	2.4999999949999952E-18
+0.35	-4.0	-1.7143528189869863E-7	2.9390055879686266E-13	0.1468584020574012	0.215673902548533	-3.999999995999997E-10	1.5999999967999978E-18
+0.39999999999999997	-3.0	-1.7143528189869863E-7	2.9390055879686266E-13	0.24685697325642747	0.6093836524532454	-2.999999996999998E-10	8.999999981999987E-19
+0.44999999999999996	-2.0	-1.7143528189869863E-7	2.9390055879686266E-13	0.3468555444554537	1.2030876871948921	-1.9999999979999986E-10	3.9999999919999946E-19
+0.49999999999999994	-1.0	-1.7143528121559864E-7	2.939005564547139E-13	0.44685411565447997	1.9967860067734735	-9.999999989999983E-11	9.999999979999965E-20
+0.5499999999999999	0.0	-3.001876656481329E-15	9.011263460727523E-29	0.5468526868535062	2.9904786111889887	-2.3453638164451093E-17	5.500731431489968E-33
+0.6	1.0	6.021537912912566E-5	3.625891883664342E-8	0.6468512580525323	4.184165500441438	0.09999000099989992	0.09998000299959987
+0.65	2.0	0.12416816168270708	0.15417732375662888	0.7468498292515586	5.577846674530822	0.19998000199980007	0.3999200119984004
+0.7000000000000001	3.0	0.22711406710624535	0.5158079947754012	0.846848400450585	7.1715221334571435	0.2999700029997003	0.899820026996402
+0.7500000000000001	4.0	0.3252570918860797	1.057921758221897	0.9468469716496113	8.965191877220398	0.3999600039996002	1.599680047993602
+0.8000000000000002	5.0	0.42388771102284667	1.7968079155618837	1.0468455428486372	10.95885590582058	0.49995000499950004	2.499500074990001
+0.8500000000000002	6.0	0.5228032898483536	2.7332327987626157	1.1468441140476637	13.152514219257707	0.5999400059994002	3.599280107985604
+0.9000000000000002	7.0	0.6219051605373214	3.8676602870295143	1.24684268524669	15.546166817531764	0.6999300069993002	4.899020146980404
+0.9500000000000003	8.0	0.7211401318783957	5.200430898055899	1.3468412564457157	18.13981370064274	0.7999200079992	6.398720191974403
+1.0000000000000002	9.0	0.8204729742565338	6.731759014853627	1.4468398276447432	20.9334548685907	0.8999100089991007	8.098380242967615
+1.0500000000000003	10.0	0.9198817194161771	8.461823777160625	1.5468383988437688	23.927090321375545	0.9999000099990003	9.998000299960008
